requirements.txt
appdirs==1.4.3
argon2-cffi==19.1.0
atomicwrites==1.3.0
attrs==19.1.0
black==19.3b0
Brotli==1.0.7
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
coverage==4.5.4
defusedxml==0.6.0
Django==2.1.12
django-admin-honeypot==1.1.0
django-allauth==0.39.1
django-cachalot==2.1.0
django-debug-toolbar==1.11
django-environ==0.4.5
django-robots==3.1.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.8
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
oauthlib==3.1.0
packaging==19.2
pluggy==0.13.0
psycopg2-binary==2.8.3
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pyparsing==2.4.2
pytest==4.6.5
pytest-django==3.5.1
pytest-sugar==0.9.2
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
sentry-sdk==0.8.0
six==1.12.0
sqlparse==0.3.0
termcolor==1.1.0
toml==0.10.0
urllib3==1.25.3
wcwidth==0.1.7
whitenoise==4.1.3
zipp==0.6.0
